+++ shared/src/constants.rs
@@ -54,6 +54,7 @@ pub const ITEM_REWARDS_MIN_LEVEL: u16 = 50;
 pub const ITEM_REWARDS_MAP_MIN_LEVEL: u16 = 100;
 pub const ITEM_REWARDS_BOSS_FACTOR: f64 = 10.0;
 pub const ITEM_REWARDS_RARE_FACTOR: f64 = 69.0;
+pub const UNIQUE_ITEM_SELL_GEMS_LEVEL_DIVISOR: f64 = 20.0;
 
 pub const STASH_USER_PRICE: StashPrice = StashPrice {
     start_price: 1e5,